Fix issue with exceptions
